+++ soil FDR/code/FDR V1.1/Inc/filter.h
@@ -6,7 +6,7 @@
 #define N 48  
 #define AN 50
 
-uint32_t filter(uint32_t seq);  
+float filter(uint32_t seq);  
 typedef struct{
   	float aN;
 	float aF;
@@ -26,7 +26,8 @@ typedef struct{
 ratio_stru *CalibrationRatio(float ahumid,float chumid);
 ratio_stru *getRatio(void);
 //uint32_t average_filter(uint32_t *pb) ;
-uint32_t average_filter(uint32_t *pb,unsigned int count);  //中值滤波+均值滤波
+float average_filter(float *pb,unsigned int count);  //中值滤波+均值滤波
 float average_filterFloat(float *pb,unsigned int sampleCount);  //中值滤波+均值滤波;
+
 #endif
